# docs/swagger.json
{"swagger":"2.0","info":{"description":"Api Documentation","version":"1.0","title":"Api Documentation","termsOfService":"urn:tos","contact":{},"license":{"name":"Apache 2.0","url":"http://www.apache.org/licenses/LICENSE-2.0"}},"host":"localhost:9000","tags":[{"name":"health-check-controller","description":"Health Check Controller"},{"name":"o-auth-2-controller","description":"O Auth 2 Controller"},{"name":"product-controller","description":"Product Controller"}],"paths":{"/gatewayapi/health/check":{"get":{"tags":["health-check-controller"],"summary":"Check microservice status","operationId":"checkUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/gatewayapi/oauth2/create":{"post":{"tags":["o-auth-2-controller"],"summary":"Creates a new client authorization","operationId":"createUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"createClientCredentialsDTO","description":"createClientCredentialsDTO","required":true,"schema":{"$ref":"#/definitions/CreateClientCredentialsDTO"}}],"responses":{"200":{"description":"Success request","schema":{"type":"object"}},"201":{"description":"Created"},"400":{"description":"Error request"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/gatewayapi/oauth2/token":{"post":{"tags":["o-auth-2-controller"],"summary":"Requests a token","operationId":"generateUsingPOST","consumes":["application/x-www-form-urlencoded"],"produces":["*/*"],"parameters":[{"name":"grant_type","in":"formData","required":false,"type":"string"},{"name":"client_secret","in":"formData","required":false,"type":"string"},{"name":"client_id","in":"formData","required":false,"type":"string"}],"responses":{"200":{"description":"Success request","schema":{"type":"object"}},"201":{"description":"Created"},"400":{"description":"Error request"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/gatewayapi/product":{"get":{"tags":["product-controller"],"summary":"Request all products","operationId":"getAllUsingGET","produces":["*/*"],"parameters":[{"name":"order","in":"query","description":"asc/desc","required":false,"type":"string","default":"desc"},{"name":"orderBy","in":"query","description":"Field name to order","required":false,"type":"string","default":"id"}],"responses":{"200":{"description":"Success request","schema":{"type":"object"}},"400":{"description":"Error request"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"post":{"tags":["product-controller"],"summary":"Create a new product","operationId":"createUsingPOST_1","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"createProductDTO","description":"createProductDTO","required":true,"schema":{"$ref":"#/definitions/CreateProductDTO"}}],"responses":{"200":{"description":"Success request","schema":{"type":"object"}},"201":{"description":"Created"},"400":{"description":"Error request"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/gatewayapi/product/byName/{name}":{"get":{"tags":["product-controller"],"summary":"Search one product searching by name","operationId":"getByNameUsingGET","produces":["*/*"],"parameters":[{"name":"name","in":"path","description":"name","required":true,"type":"string"}],"responses":{"200":{"description":"Success request","schema":{"type":"object"}},"400":{"description":"Error request"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/gatewayapi/product/{id}":{"get":{"tags":["product-controller"],"summary":"Request one product searching by id","operationId":"getByIdUsingGET","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"string"}],"responses":{"200":{"description":"Success request","schema":{"type":"object"}},"400":{"description":"Error request"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"put":{"tags":["product-controller"],"summary":"Update a product by its id","operationId":"updateUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"createProductDTO","description":"createProductDTO","required":true,"schema":{"$ref":"#/definitions/CreateProductDTO"}},{"name":"id","in":"path","description":"id","required":true,"type":"string"}],"responses":{"200":{"description":"Success request","schema":{"type":"object"}},"201":{"description":"Created"},"400":{"description":"Error request"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"delete":{"tags":["product-controller"],"summary":"Remove a product by its id","operationId":"deleteByIdUsingDELETE","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"string"}],"responses":{"200":{"description":"Success request","schema":{"type":"object"}},"204":{"description":"No Content"},"400":{"description":"Error request"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}}},"definitions":{"ClientCredentialsTO":{"type":"object","properties":{"client_id":{"type":"string"},"client_secret":{"type":"string"},"grant_type":{"type":"string"}},"title":"ClientCredentialsTO"},"CreateClientCredentialsDTO":{"type":"object","properties":{"client_secret":{"type":"string"}},"title":"CreateClientCredentialsDTO"},"CreateProductDTO":{"type":"object","properties":{"description":{"type":"string"},"name":{"type":"string"},"price":{"type":"string"},"type":{"type":"string"}},"title":"CreateProductDTO"},"ErrorDTO":{"type":"object","properties":{"code":{"type":"string"},"message":{"type":"string"}},"title":"ErrorDTO"},"IdDTO":{"type":"object","properties":{"id":{"type":"string"}},"title":"IdDTO"},"ProductTO":{"type":"object","properties":{"description":{"type":"string"},"id":{"type":"string"},"name":{"type":"string"},"price":{"type":"string"},"type":{"type":"string"}},"title":"ProductTO"},"TokenTO":{"type":"object","properties":{"expires":{"type":"integer","format":"int64"},"token":{"type":"string"}},"title":"TokenTO"}}}